% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Regression.R
\name{adaptive.regress.stabs}
\alias{adaptive.regress.stabs}
\title{Generate adaptive regression results for gene-enhancer association}
\usage{
adaptive.regress.stabs(
  object,
  rep3_data,
  PFER,
  q,
  peak.assay = "ATAC",
  expression.assay = "SCT",
  expression.slot = "data",
  peak.slot = "counts"
)
}
\arguments{
\item{object}{a data frame has 3 columns: gene (gene name), peak (matched cis-peak), coaccess (coaccessibility calculated by cicero between the peak and the promoter of the gene)}

\item{rep3_data}{a seurat object}

\item{PFER}{upper bound per family error rate}

\item{q}{number of (unique) selected variables (or groups of variables depending on the model) that are selected on each subsample.}

\item{peak.assay}{the name of the assay containing enhancer accessibility}

\item{expression.assay}{the name of the assay containing gene expression}

\item{expression.slot}{the name of the slot to use in expression assay}

\item{peak.slot}{the name of the slot to use in enhancer accessibility assay}
}
\value{
a data frame with adaptive regression results: gene (gene name), peak (paired cis-peak to this gene), max_rate (adaptive regression result for the gene-peak pair)
}
\description{
Generate adaptive regression results for gene-enhancer association
}
