% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coaccess.adapt.data.K562}
\alias{coaccess.adapt.data.K562}
\title{Coaccessibility from cicero for running a demo}
\format{
a dataset with 510 rows and 3 columns:
\describe{
  \item{Peak1}{First peak input for cicero}
  \item{Peak2}{Second peak input for cicero}
  \item{gene}{gene name}
  \item{peak}{cis-peak region}
  \item{coaccess}{coaccessibility computed by cicero between the gene promoter and peak}
  \item{ind}{indicator to be 1 or 2 indicating which peak before is the promoter of the gene}
}
}
\usage{
coaccess.adapt.data.K562
}
\description{
Coaccessibility from cicero for running a demo
}
\keyword{datasets}
