% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Distance_adjust}
\alias{Distance_adjust}
\title{Distance adjusted statistics}
\usage{
Distance_adjust(result, gene.coords)
}
\arguments{
\item{result}{result statistics without distance adjustment}

\item{gene.coords}{data frame for gene coordinates}
}
\value{
data frame with one additional column named distance_adjusted for distance adjusted statistics
}
\description{
Distance adjusted statistics
}
