% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEEES.R
\name{edge}
\alias{edge}
\title{Compute cell-specific GEEES statistics for a specific gene-enhancer pair}
\usage{
edge(peak_data, gene_data, num)
}
\arguments{
\item{peak_data}{a peak accessibility vector.}

\item{gene_data}{a gene expression vector.}

\item{num}{size of the cell neighbor.}
}
\value{
a vector containing GEEES statistics for gene-enhancer pair on each single cell
}
\description{
Compute cell-specific GEEES statistics for a specific gene-enhancer pair
}
