% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_peak}
\alias{find_peak}
\title{Find overlap between two lists of chromatin regions}
\usage{
find_peak(peaks, genes, sep = c("-", "-"))
}
\arguments{
\item{peaks}{first list of chromatin regions}

\item{genes}{second list of chromatin regions}

\item{sep}{the vector of separator to use for genome string}
}
\value{
a binary matrix indicating if two pairs of peaks are overlapped with first list on the row and second list on the column
}
\description{
Find overlap between two lists of chromatin regions
}
\keyword{internal}
