% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEEES.R
\name{generate_pair}
\alias{generate_pair}
\title{Aggregating cell-specific GEEES statistic over the population for a summarized statistic}
\usage{
generate_pair(data)
}
\arguments{
\item{data}{a seurat object}
}
\value{
a data frame with aggregated GEEES statistics for all gene-enhancer pairs

a data frame with 3 columns: gene (gene name), peak (the paired cis-peak to the gene), max_rate (aggregated GEEES statistic for this pair)
}
\description{
Aggregating cell-specific GEEES statistic over the population for a summarized statistic
}
