% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{K562_mini}
\alias{K562_mini}
\title{A list containing normalized gene expression and peak accessibility for K562 multi-modal single-cell data with 200 cells and 5 genes}
\description{
A list containing normalized gene expression (in SCT) and peak accessibility (in ATAC) for K562 multi-modal single-cell data with 200 cells and 5 genes.
}
