% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEEES.R
\name{NetworkInCell}
\alias{NetworkInCell}
\title{Compute cell-specific GEEES statistics for all specified gene with their cis-peaks}
\usage{
NetworkInCell(
  object,
  peak.assay,
  expression.assay,
  expression.slot = "data",
  peak.slot = "counts",
  distance = 5e+05,
  min.distance = NULL,
  min.cells = 1,
  genes.use = NULL,
  verbose = T,
  cl,
  num
)
}
\arguments{
\item{object}{a Seurat object}

\item{peak.assay}{the name of the assay containing enhancer accessibility}

\item{expression.assay}{the name of the assay containing gene expression}

\item{expression.slot}{the name of the slot to use in expression assay}

\item{peak.slot}{the name of the slot to use in enhancer accessibility assay}

\item{distance}{the window size of peak considered as cis-peak for genes}

\item{min.distance}{the minimum distance between gene and peak for a peak to be considered as a candidate enhancer}

\item{min.cells}{gene and peak filtering where only genes and peaks that are expressed/accessible in at least min.cells cells are considered}

\item{genes.use}{a vector of gene names for enhancer detection (all genes in the object will be used if not given)}

\item{verbose}{if the total numbers of genes and peaks are reported}

\item{cl}{a list of clusters for parallel computing}

\item{num}{number of cells in a cell neighbor}
}
\value{
a list of matrices where each element is a cell by peak matrix containing the cell-specific GEEES statistics for a specific gene and its cis-peaks
}
\description{
Compute cell-specific GEEES statistics for all specified gene with their cis-peaks
}
