% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Regression.R
\name{rem.function}
\alias{rem.function}
\title{multi-response regression for one gene}
\usage{
rem.function(x, y, q, ...)
}
\arguments{
\item{x}{peak accessibility matrix}

\item{y}{gene expression vector}

\item{q}{the number of selected peaks}

\item{...}{other arguments can be passed}
}
\value{
a list of a logical vector indicating which peaks are selected at this round of stability selection
}
\description{
multi-response regression for one gene
}
