% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Regression.R
\name{remMap.regression.stabs}
\alias{remMap.regression.stabs}
\title{Generate multi-response regression results for gene-enhancer association}
\usage{
remMap.regression.stabs(
  object,
  rep3,
  peak.assay = "ATAC",
  expression.assay = "RNA",
  expression.slot = "data",
  peak.slot = "counts",
  PFER,
  q
)
}
\arguments{
\item{object}{a data frame has 2 columns: gene (gene name), peak (matched cis-peak)}

\item{rep3}{a seurat object}

\item{peak.assay}{the name of the assay containing enhancer accessibility}

\item{expression.assay}{the name of the assay containing gene expression}

\item{expression.slot}{the name of the slot to use in expression assay}

\item{peak.slot}{the name of the slot to use in enhancer accessibility assay}

\item{PFER}{upper bound per family error rate}

\item{q}{number of (unique) selected variables (or groups of variables depending on the model) that are selected on each subsample.}
}
\value{
a data frame with adaptive regression results: gene (gene name), peak (paired cis-peak to this gene), max_rate (multi-response regression result for the gene-peak pair)
}
\description{
Generate multi-response regression results for gene-enhancer association
}
