% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Regression.R
\name{sequential.regression.stabs}
\alias{sequential.regression.stabs}
\title{Generate sequential regression results for gene-enhancer association}
\usage{
sequential.regression.stabs(
  object,
  rep3_data,
  peak.assay = "ATAC",
  expression.assay = "RNA",
  expression.slot = "data",
  peak.slot = "counts",
  PFER,
  q
)
}
\arguments{
\item{object}{a data frame has 5 columns: Peak1 (the first peak as input for cicero), Peak2 (the second peak as input for cicero), gene (gene name), peak (matched cis-peak), ind (indicator to be 1 or 2 indicating which peak before is the promoter for the gene)}

\item{rep3_data}{a seurat object}

\item{peak.assay}{the name of the assay containing enhancer accessibility}

\item{expression.assay}{the name of the assay containing gene expression}

\item{expression.slot}{the name of the slot to use in expression assay}

\item{peak.slot}{the name of the slot to use in enhancer accessibility assay}

\item{PFER}{upper bound per family error rate}

\item{q}{number of (unique) selected variables (or groups of variables depending on the model) that are selected on each subsample.}
}
\value{
a data frame with adaptive regression results: gene (gene name), peak (paired cis-peak to this gene), max_rate (sequential regression result for the gene-peak pair)
}
\description{
Generate sequential regression results for gene-enhancer association
}
